Plugins CLI and API (#404)
* Plugin CLI and API implementation.  Closes #396.

First pass at plugin installation CLI.

Using correct types in YAML options, adding line breaks between newly added sections.

Fixing issues serializing YAML changes. Adding support for rewriting JSON.

Moving plugin installation logic to lib/plugin-installer.js

Fixing linter issues.

Updating Swagger doc with plugin definitions.

Updating swagger-ui-express and firming up Plugin API design.

Fixing linter issues.

Adding REST API for Plugins.

Finishing plugin API implementation.

Adding `eg plugin configure` generator.

Making system config for plugins always have a package property. Gets around YAWN's empty object issue.

Fixing plugin name resolution, allowing 'eg plugin configure example'.

Removing double-blank line from bin/generators/plugin/configure.js

* Adding plugin installer tests.

* Adding plugin install e2e test.

* Adding plugin CLI e2e test.

* Minor changes to plugin installer.

* Allowing eg plugin install to work on npm@4 and npm@5. Fixing stray node-uuid references in tests.

* Fixing borked key-auth e2e test.

* Hiding Plugins API behind feature flag --experimental-plugins-api.

* Style fixes, putting plugins api behind --experimental-plugins-api flag, making plugin-installer async.

* Fixing some linter issues, plugin-installer bug, and making a friendlier error response on invalid plugin configuration via the CLI.

+++ test/e2e/key-auth.e2e.test.js
@@ -1,6 +1,7 @@
+const request = require('supertest');
 const cliHelper = require('../common/cli.helper');
 const gwHelper = require('../common/gateway.helper');
-const request = require('supertest');
+
 let gatewayProcess = null;
 let gatewayPort, adminPort, configDirectoryPath;
 let username = 'test';
@@ -99,7 +100,10 @@ describe('E2E: key-auth Policy', () => {
   });
 
   after('cleanup', (done) => {
-    gatewayProcess.kill();
+    if (gatewayProcess) {
+      gatewayProcess.kill();
+    }
+
     done();
   });
